Optimize Indicator params (pf=1.05; dd=2.97; tt=7144)

+++ config/EURUSD_M15.h
@@ -18,7 +18,7 @@ struct Stg_DeMarker_Params_M15 : StgParams {
     lot_size = 0;
     signal_open_method = 0;
     signal_open_filter = 1;
-    signal_open_level = (float)0.2;
+    signal_open_level = (float)(float)0.2;
     signal_open_boost = 0;
     signal_close_method = 0;
     signal_close_level = (float)0;
